Worked on reviewing store

+++ lib/features/core/models/product_model.dart
@@ -110,7 +110,7 @@ class Store {
   String location;
   double lon;
   double lat;
-  int star;
+  double star;
   String logo;
 
   factory Store.fromJson(Map<String, dynamic> json) => Store(
@@ -119,7 +119,7 @@ class Store {
         location: json["location"],
         lon: json["lon"].toDouble(),
         lat: json["lat"].toDouble(),
-        star: json["star"],
+        star: json["star"].toDouble(),
         logo: json["logo"],
       );
